Move WebSocketClosed to errors.py, rename it to WebSocketDisconnected, get rid of checking for disconnection in web.RequestHandler
