+++ tests/test_web_websocket.py
@@ -2,7 +2,7 @@ import asyncio
 import unittest
 from unittest import mock
 from aiohttp.multidict import MultiDict
-from aiohttp.web import (Request, WebSocketResponse, WebSocketClosed,
+from aiohttp.web import (Request, WebSocketResponse, WebSocketDisconnected,
                          HTTPMethodNotAllowed, HTTPBadRequest)
 from aiohttp.protocol import RawRequestMessage, HttpVersion11
 
@@ -90,7 +90,7 @@ class TestWebWebSocket(unittest.TestCase):
 
         @asyncio.coroutine
         def a():
-            raise WebSocketClosed()
+            raise WebSocketDisconnected()
 
         @asyncio.coroutine
         def b():
@@ -100,7 +100,7 @@ class TestWebWebSocket(unittest.TestCase):
         def c():
             yield from b()
 
-        with self.assertRaises(WebSocketClosed):
+        with self.assertRaises(WebSocketDisconnected):
             self.loop.run_until_complete(c())
 
     def test_can_start_ok(self):
